+++ vNugglets.Utility/vNuggletsUtilityMod.psm1
@@ -1,6 +1,3 @@
-## dot-source the config items file
-#. $PSScriptRoot\configItems.ps1
-
 ## Set aliases; when setting value for DefaultCommandPrefix in module, need to account for that when setting value for Alias here (value needs to be of what the final function name will be with that DefaultCommandPrefix)
 $hshNewAliasInfo = @{
 	## Connect and Disconnect VIServer
@@ -13,14 +10,11 @@ $arrAliasNamesToExport = $hshNewAliasInfo.GetEnumerator() | Foreach-Object {
 	$strNameThisAlias,"VN$strNameThisAlias" | Foreach-Object {if (-not (Get-Alias -Name $_ -ErrorAction:SilentlyContinue)) {New-Alias -Name $_ -Value $strValueThisAlias; $_}}
 } ## end foreach-object
 
-## export these items for use by consumer
-$hshModuleMemberParams = @{
-	Function = Write-Output Connect-VNVIServer, Copy-VNVIRole, Disconnect-VNVIServer, Find-VNVMWithDuplicateMACAddress,
-		Get-VNNetworkClusterInfo, Get-VNUplinkNicForVM, Get-VNVMByAddress, Get-VNVMByRDM, Get-VNVMByVirtualPortGroup, Get-VNVMDiskAndRDM, Get-VNVMEVCInfo,
-		Get-VNVMHostBrokenUplink, Get-VNVMHostFirmwareInfo, Get-VNVMHostHBAWWN, Get-VNVMHostLogicalVolumeInfo, Get-VNVMHostNICFirmwareAndDriverInfo,
-		Invoke-VNEvacuateDatastore, Move-VNTemplateFromVMHost, Update-VNTitleBarForPowerCLI
-	Alias = $arrAliasNamesToExport
-} ## end hsh
+## now handled in module manifest
+# ## export these items for use by consumer
+# $hshModuleMemberParams = @{
+# 	Alias = $arrAliasNamesToExport
+# } ## end hsh
 
-## do the actual member export
-Export-ModuleMember @hshModuleMemberParams
+# ## do the actual member export
+# Export-ModuleMember @hshModuleMemberParams