+++ ChangeLog.md
@@ -1,6 +1,6 @@
 ### ChangeLog for vNugglets.Utility PowerShell module
 
-#### v1.next, not yet released
+#### v1.2, not yet released
 
 - \[improvement] updated function `Get-VNVMByAddress`:
   - greatly increased speed by using `VMware.Vim.SearchIndex` for by-IP searches (does support wildcard, though -- by wildcard still uses slower mechanism); example speed increase: went from 13s to about 0.25s in a vCenter with about 7,500 VMs -- woo-hoo!
@@ -9,6 +9,7 @@
   - takes new parameter, `-SourceRole`, for passing VIRole object itself as value, and this parameter accepts value from pipeline, for more natural use of cmdlet
   - deduces source vCenter server from `-SourceRole` value, simplifying use of cmdlet (no need to specify `-SourceVCName` parameter when providing the source VIRole object)
   - `-DestinationVCName` parameter now optional, further simplifying use of cmdlet; if parameter not specified, destination vCenter will be the same as the source vCenter
+- \[internal improvement] updated to using `Update-ModuleManifest` for keeping module manifest in shape, added manifest entries for tags and for URIs for project, release notes, license, etc. (partially in preparation for publishing to the [PowerShellGallery](https://www.powershellgallery.com/))
   
 
 #### v1.1, released 20 Dec 2016